# docs/openapi.yaml
openapi: 3.0.1
info:
  title: ldap2azure
  description: 'The following describes the API of ldap2azure. The API is used by the frontend, but can also be used with API key authentication from a backend, e.g. for monitoring.'
  license:
    name: GPL-3.0 License
    url: https://www.gnu.org/licenses/gpl-3.0.en.html
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
  - url: https://localhost:8080/api
...
paths:
...
  /user:
    get:
      tags:
        - user
      summary: Retrieve the users currently stored in the database.
      description: This endpoint retreives the users synced by ldap2azure. Since the number of users in the system can be very high, depending on the size of the environment, the end point has a multi-page structure and only returns a maximum of 20 objects at a time.
      operationId: userGetUsers
      parameters:
        - name: page
          schema:
            type: integer
            default: 0
            example: 2
          in: query
          required: false
      responses:
        200:
          description: Request successfull. The maximum ammount of 20 objects of the given page is returned.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        401:
          description: You are not authenticated or your authentication secret is invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleLineError'
      security:
        - cookieAuth: []
        - api_key: []
  /user/ok:
    get:
      tags:
        - user
      summary: Retrieve the users that currently have the OK sync status.
      description: Since the number of users in the system can be very high, depending on the size of the environment, the end point has a multi-page structure and only returns a maximum of 20 objects at a time.
      operationId: userGetUsersOk
      parameters:
        - name: page
          schema:
            type: integer
            default: 0
            example: 2
          in: query
          required: false
      responses:
        200:
          description: Request successfull. The maximum ammount of 20 objects of the given page is returned.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        401:
          description: You are not authenticated or your authentication secret is invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleLineError'
      security:
        - cookieAuth: []
        - api_key: []
  /user/pending:
    get:
      tags:
        - user
      summary: Retrieve the users that currently have the PENDING sync status.
      description: Since the number of users in the system can be very high, depending on the size of the environment, the end point has a multi-page structure and only returns a maximum of 20 objects at a time.
      operationId: userGetUsersPending
      parameters:
        - name: page
          schema:
            type: integer
            default: 0
            example: 2
          in: query
          required: false
      responses:
        200:
          description: Request successfull. The maximum ammount of 20 objects of the given page is returned.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        401:
          description: You are not authenticated or your authentication secret is invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleLineError'
      security:
        - cookieAuth: []
        - api_key: []
  /user/failed:
    get:
      tags:
        - user
      summary: Retrieve the users that currently have the FAILED sync status.
      description: Since the number of users in the system can be very high, depending on the size of the environment, the end point has a multi-page structure and only returns a maximum of 20 objects at a time.
      operationId: userGetUsersFailed
      parameters:
        - name: page
          schema:
            type: integer
            default: 0
            example: 2
          in: query
          required: false
      responses:
        200:
          description: Request successfull. The maximum ammount of 20 objects of the given page is returned.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        401:
          description: You are not authenticated or your authentication secret is invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleLineError'
      security:
        - cookieAuth: []
        - api_key: []
  /user/{id}:
    get:
      tags:
        - user
      summary: Get a single user object from the database.
      operationId: userGetUser
      parameters:
        - name: id
          description: Id of the user that should be retreived.
          schema:
            type: string
            example: cb78467f-6dbd-4563-ac01-0fdc9ca4c194
          in: path
          required: true
      responses:
        200:
          description: Request successfull. The user with the given id is returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        401:
          description: You are not authenticated or your authentication secret is invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleLineError'
        404:
          description: Not found. The user with the given id not existing in the database.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleLineError'
              example:
                error: user_not_existing
      security:
        - cookieAuth: []
        - api_key: []
  /user/{id}/retry:
    post:
      tags:
        - user
      summary: Retry the sync to Azure AD for the given user.
      operationId: userRetrySync
      parameters:
        - name: id
          description: Id of the user that should be retried.
          schema:
            type: string
            example: cb78467f-6dbd-4563-ac01-0fdc9ca4c194
          in: path
          required: true
      responses:
        200:
          description: The retry was successfull. The user object is returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        401:
          description: You are not authenticated or your authentication secret is invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleLineError'
        403:
          description: You have insufficient permissions. This results from an read-only permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleLineError'
              example:
                error: "insufficient_permissions"
        404:
          description: Not found. The user with the given id not existing in the database.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleLineError'
              example:
                error: user_not_existing
        500:
          description: An error occured during the retry. The error message holds further information.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TwoLineError'
      security:
        - cookieAuth: []
        - api_key: []
  /user/status:
    get:
      tags:
        - user
      summary: Get a status of how many users are in each sync status. Useful for monitoring tools.
      operationId: userGetStatus
      responses:
        200:
          description: The request was sucessfull. The ammount of users in each sync status is returned.
          content:
            application/json:
              schema:
                type: object
                properties:
                  userCount:
                    type: integer
                    description: Overall ammount of users known by ldap2azure.
                  usersOk:
                    type: integer
                    description: Ammount of users that have the OK sync state.
                  usersPending:
                    type: integer
                    description: Number of users waiting for a sync to Azure AD.
                  usersFailed:
                    type: integer
                    description: Number of users that failed to sync.
        401:
          description: You are not authenticated or your authentication secret is invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleLineError'
      security:
        - cookieAuth: []
        - api_key: []
  /user/{id}/conflicts:
    get:
      tags:
        - user
      summary: Get possible conflicts for the internal user with the given id.
      operationId: userGetConflicts
      parameters:
        - name: id
          description: Id of the user for which possible conflicts should be retrieved for.
          schema:
            type: string
            example: cb78467f-6dbd-4563-ac01-0fdc9ca4c194
          in: path
          required: true
      responses:
        200:
          description: The request was sucessfull. Possible conflicting users in Azure are returned as an array. The changeState property of the returned users indicates where the users are located in Azure AD. They can be located in the trashbin (changeState == deleted) or be still active user accounts (changeState == unchanged). Not alle properties of the User object are populated. See the below example.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
                example:
                  - onPremisesImmutableId: 72ea6SbbD0WSr3iT3oCrkQ==
                    azureImmutableId: cb78467f-6dbd-4563-ac01-0fdc9ca4c194
                    givenName: Max
                    surname: Mustermann
                    displayName: Max Mustermann
                    userPrincipalName: Max.Mustermann@example.com
                    hash: 2VQreuFtzfPKGqHeVhHqBw==
                    changeState: unchanged
                    lastChanged: '2020-08-13T21:35:33.140Z'
        400:
          description: The given user is not in the failed state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleLineError'
              example:
                error: user_not_failed
        401:
          description: You are not authenticated or your authentication secret is invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleLineError'
        403:
          description: You have insufficient permissions. This results from an read-only permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleLineError'
              example:
                error: "insufficient_permissions"
        404:
          description: Not found. The user with the given id not existing in the database.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleLineError'
              example:
                error: user_not_existing
      security:
        - cookieAuth: []
        - api_key: []
  /user/{id}/conflicts/resolve:
    post:
      tags:
        - user
      summary: Resolve an user conflict by merging or recreating user in Azure AD.
      operationId: userResolveConflict
      parameters:
        - name: id
          description: Id of the user for which an conflict should be resolved.
          schema:
            type: string
            example: cb78467f-6dbd-4563-ac01-0fdc9ca4c194
          in: path
          required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConflictResolveRequest'
      responses:
        200:
          description: The request was sucessfull. The updated User object is returned.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        400:
          description: One of the two users is in a wrong status.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/SingleLineError'
                  - $ref: '#/components/schemas/TwoLineError'
              examples:
                user_not_failed:
                  summary: The given (internal) user is not in the failed state.
                  value:
                    error: user_not_failed
                azureId_already_assigned:
                  summary: The given Azure user id is already asigned to another user in ldap2azure.
                  value:
                    error: azureId_already_assigned
                    message: The provided Azure user id is already assigned to the internal user with id cb78467f-6dbd-4563-ac01-0fdc9ca4c194
        401:
          description: You are not authenticated or your authentication secret is invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleLineError'
        403:
          description: You have insufficient permissions. This results from an read-only permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleLineError'
              example:
                error: "insufficient_permissions"
        404:
          description: Not found. The user with the given id not existing in the database.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleLineError'
              example:
                error: user_not_existing
        500:
          description: An error has occurred during communication with Azure. The error message holds further information.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TwoLineError'
      security:
        - cookieAuth: []
        - api_key: []
...
components:
  schemas:
...
    User:
      type: object
      properties:
        _id:
          type: string
          example: cb78467f-6dbd-4563-ac01-0fdc9ca4c194
        onPremisesImmutableId:
          type: string
          example: 72ea6SbbD0WSr3iT3oCrkQ==
        azureImmutableId:
          type: string
          example: cb78467f-6dbd-4563-ac01-0fdc9ca4c194
        givenName:
          type: string
          example: Max
        surname:
          type: string
          example: Mustermann
        displayName:
          type: string
          example: Max Mustermann
        mailNickname:
          type: string
          example: max.mustermann
        userPrincipalName:
          type: string
          example: Max.Mustermann@example.com
        hash:
          type: string
          example: 2VQreuFtzfPKGqHeVhHqBw==
        syncState:
          type: string
          enum:
            - "ok"
            - "pending"
            - "failed"
          example: ok
        changeState:
          type: string
          enum:
            - new
            - unchanged
            - changed
            - deleted
          example: unchanged
        lastChanged:
          type: string
          format: date-time
        lastSyncId:
          type: string
          example: cb78467f-6dbd-4563-ac01-0fdc9ca4c194
...
    ConflictResolveRequest:
      type: object
      properties:
        azureImmutableId:
          type: string
          example: cb78467f-6dbd-4563-ac01-0fdc9ca4c194
          description: Id of the azure user the conflict should be resolved with.
        strategy:
          type: string
          enum: [merge, recreate]
          example: merge
          description: Use merge to merge both users, or recreate to delete the azure user completeley including all its data and recreate it from scratch.
    SingleLineError:
      type: object
      properties:
        error:
          type: string
          example: not_authenticated
    TwoLineError:
      type: object
      properties:
        error:
          type: string
          example: error_from_azure
        message:
          type: string
          example: Another object with the same value for property userPrincipalName already exists.
  securitySchemes:
    api_key:
      type: apiKey
      name: X-API-Key
      in: header
    cookieAuth:
      type: apiKey
      in: cookie
      name: cdsess